module9-task2: * event price component
