+++ src/helpers/trip/index.js
@@ -1,5 +1,4 @@
 export * from './get-unique-days/get-unique-days.helper';
-export * from './get-total-price/get-total-price.helper';
 export * from './get-sorted-events-by-price/get-sorted-events-by-price';
 export * from './get-sorted-events-by-duration/get-sorted-events-by-duration.helper';
 export * from './get-trip-offers-by-type/get-trip-offers-by-type.helper';
