+++ src/presenter/destination-info/helpers.js
@@ -1,5 +1,7 @@
 import {getUniqueItems} from '~/helpers';
 
+const DEFAULT_PRICE = 0;
+
 const getUniqueEventCities = (events) => {
   const cities = events.map((it) => it.destination.city);
   const uniqueCities = getUniqueItems(cities);
@@ -7,4 +9,18 @@ const getUniqueEventCities = (events) => {
   return uniqueCities;
 };
 
-export {getUniqueEventCities};
+const getTotalPrice = (events) => {
+  const totalPrice = events.reduce((totalPriceAcc, event) => {
+    const offersPrice = event.offers.reduce(
+        (offersPriceAcc, offer) =>
+          (offersPriceAcc += offer.isChecked ? offer.price : DEFAULT_PRICE),
+        DEFAULT_PRICE
+    );
+
+    return (totalPriceAcc += (event.price + offersPrice));
+  }, DEFAULT_PRICE);
+
+  return totalPrice;
+};
+
+export {getUniqueEventCities, getTotalPrice};
