+++ src/view/trip-price/trip-price.js
@@ -1,7 +1,7 @@
 import Abstract from '~/view/abstract/abstract';
 
 class DestinationInfo extends Abstract {
-  constructor(price) {
+  constructor({price}) {
     super();
     this.price = price;
   }
